Add functions to get system OS and arch
